Small Changes
1. Even when changing the hand to the front and back, the color should be changed and the text displayed or hidden.
2.When the hand color is blue, it means that the hand is accurately tracked.
In this case, the character is displayed on the finger and the character is input according to the movement of the finger.

However, when the hand color is green, it means that the hand was not tracked correctly.
In this case, the character should not be displayed on the finger, and the character should not be input even if the finger is moved.
That is, only when a character is displayed on the finger, a character related to each finger can be input.